event model refactor
