+++ gql_events/DBDefinitions/base.py
@@ -1,0 +1,4 @@
+from sqlalchemy.orm import DeclarativeBase
+
+class BaseModel(DeclarativeBase):
+    pass
