+++ gql_events/DBFeeder.py
@@ -1,7 +1,8 @@
 from functools import cache
 
-from gql_events.DBDefinitions import (
-    EventModel, 
+
+from gql_events.DBDefinitions import EventModel
+from gql_events.DBDefinitionsOld import (
     EventTypeModel, 
     EventGroupModel, 
     PresenceModel,
